refactor sidebar in extra component file, update sidebar when other id
