+++ src/App.vue
@@ -1,106 +1,11 @@
 <script setup lang="ts">
-import { defineProps, defineEmits, ref, watch } from "vue";
-import { SidebarHeaderItem, SidebarItem } from "vue-sidebar-menu";
-import "vue-sidebar-menu/dist/vue-sidebar-menu.css";
-import { exampleWorlds, ICourse, IWorld } from "./ts/models";
-import { store } from "@/store";
-
-import { computed } from "vue";
-import { useRouter } from "vue-router";
-import { getCourse } from "./ts/course-rest-client";
-
-const router = useRouter();
-let course: ICourse;
-const menu = ref([]);
-
-function loadCourse(id: number) {
-  getCourse(id)
-    .then((response) => {
-      const result: ICourse = response.data;
-      course = result;
-      console.log(course);
-      loadMenu();
-    })
-    .catch((error) => {
-      console.log(error);
-    });
-}
-
-function loadMenu() {
-  if (course == null) {
-    return;
-  }
-  menu.value = [
-    {
-      header: "Main Navigation",
-      hiddenOnCollapse: false,
-    },
-  ];
-  menu.value.push({
-    href: "/courses/" + course.id,
-    title: "course : " + course.courseName,
-    icon: "bi-map-fill",
-  });
-  const worlds = course.worlds.sort(
-    (world1, world2) => world1.index + world2.index
-  );
-  worlds.forEach((world: IWorld) => {
-    menu.value.push({
-      href: "/courses/" + course.id + "/worlds/" + world.index,
-      title: world.staticName,
-      icon: "bi-map-fill",
-    });
-  });
-  console.log("LOADED MENU");
-  console.log(menu.value);
-}
-
-showSideBar().then((show) => {
-  if (show) {
-    getCourseId().then((courseId) => {
-      loadCourse(courseId);
-    });
-  }
-});
-
-const sidebarActive = ["course", "worlds"];
-
-async function showSideBar() {
-  await router.isReady();
-  const route = router.currentRoute.value.name;
-  if (route) {
-    return sidebarActive.includes(route.toString());
-  }
-  return false;
-}
-
-const showSideBarComputed = computed(() => {
-  const route = router.currentRoute.value.name;
-  if (route) {
-    return sidebarActive.includes(route.toString());
-  }
-  return false;
-});
-
-async function getCourseId(): Promise<number> {
-  console.log(showSideBar);
-  await router.isReady();
-  const args = router.currentRoute.value.fullPath.split("/");
-  console.log("ARGS");
-  console.log(args);
-  if (args.length > 2) {
-    if (args[1].toLowerCase() == "courses") {
-      return parseInt(args[2] as string);
-    }
-  }
-  return 0;
-}
+import CourseSidebarMenu from "@/components/CourseSidebarMenu.vue";
 </script>
 
 <template>
   {{ course }}
   <div class="app-wrapper">
-    <sidebar-menu :menu="menu" :relative="true" v-if="showSideBarComputed" />
+    <CourseSidebarMenu />
     <!-- route outlet -->
     <!-- component matched by the route will render here -->
     <div class="router-view-wrapper">
@@ -108,19 +13,3 @@ async function getCourseId(): Promise<number> {
     </div>
   </div>
 </template>
-
-<style>
-.app-wrapper {
-  height: 100vh;
-  display: flex;
-}
-
-.v-sidebar-menu {
-  height: 100%;
-}
-
-.router-view-wrapper {
-  width: 100%;
-  overflow: scroll;
-}
-</style>
